# src/IO.Swagger/Models/Subject.cs
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is an opeapi specification for lab project 1  Made by _Oleksandr Myroshnychenko_  Group - 6.04.121.10.19.2
 *
 * OpenAPI spec version: 1.0.11
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class Subject : IEquatable<Subject>
    { 
        /// <summary>
        /// Gets or Sets Id
        /// </summary>

        [DataMember(Name="Id")]
        public int? Id { get; set; }

        /// <summary>
        /// Gets or Sets Name
        /// </summary>

        [DataMember(Name="Name")]
        public string Name { get; set; }

        /// <summary>
        /// Gets or Sets LectureHours
        /// </summary>

        [DataMember(Name="LectureHours")]
        public int? LectureHours { get; set; }

        /// <summary>
        /// Gets or Sets PracticeHours
        /// </summary>

        [DataMember(Name="PracticeHours")]
        public int? PracticeHours { get; set; }

        /// <summary>
        /// Gets or Sets LabHours
        /// </summary>

        [DataMember(Name="LabHours")]
        public int? LabHours { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Subject {\n");
            sb.Append("  Id: ").Append(Id).Append("\n");
            sb.Append("  Name: ").Append(Name).Append("\n");
            sb.Append("  LectureHours: ").Append(LectureHours).Append("\n");
            sb.Append("  PracticeHours: ").Append(PracticeHours).Append("\n");
            sb.Append("  LabHours: ").Append(LabHours).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Subject)obj);
        }

        /// <summary>
        /// Returns true if Subject instances are equal
        /// </summary>
        /// <param name="other">Instance of Subject to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Subject other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Id == other.Id ||
                    Id != null &&
                    Id.Equals(other.Id)
                ) && 
                (
                    Name == other.Name ||
                    Name != null &&
                    Name.Equals(other.Name)
                ) && 
                (
                    LectureHours == other.LectureHours ||
                    LectureHours != null &&
                    LectureHours.Equals(other.LectureHours)
                ) && 
                (
                    PracticeHours == other.PracticeHours ||
                    PracticeHours != null &&
                    PracticeHours.Equals(other.PracticeHours)
                ) && 
                (
                    LabHours == other.LabHours ||
                    LabHours != null &&
                    LabHours.Equals(other.LabHours)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Id != null)
                    hashCode = hashCode * 59 + Id.GetHashCode();
                    if (Name != null)
                    hashCode = hashCode * 59 + Name.GetHashCode();
                    if (LectureHours != null)
                    hashCode = hashCode * 59 + LectureHours.GetHashCode();
                    if (PracticeHours != null)
                    hashCode = hashCode * 59 + PracticeHours.GetHashCode();
                    if (LabHours != null)
                    hashCode = hashCode * 59 + LabHours.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Subject left, Subject right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Subject left, Subject right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}