# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Swagger Petstore - OpenAPI 3.0",
    "description" : "This is an opeapi specification for lab project 1\n\nMade by _Oleksandr Myroshnychenko_\n\nGroup - 6.04.121.10.19.2",
    "version" : "1.0.11"
  },
  "servers" : [ {
    "url" : "https://localhost:7112/"
  } ],
...
  "paths" : {
...
    "/api/subject" : {
      "get" : {
        "tags" : [ "subject" ],
        "summary" : "Returns all subjects from db",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Subject"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "subject" ],
        "summary" : "Update an existing subject",
        "description" : "Updates an existing subject",
        "requestBody" : {
          "description" : "Update an existent subject",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Subject"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subject"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "subject" ],
        "summary" : "Create a subject in db",
        "parameters" : [ {
          "name" : "entityData",
          "in" : "header",
          "description" : "Data that will be added to db as a subject",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/entityData_1"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Entity was successfully added",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Subject"
                }
              }
            }
          }
        }
      }
    },
    "/api/subject/{id}" : {
      "get" : {
        "tags" : [ "subject" ],
        "summary" : "Finds subject by id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of a subject that needs to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "example" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Subject"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "subject" ],
        "summary" : "Deletes a subject",
        "description" : "delete a subject",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of a subject that need to be deleted",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "example" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Student was successfully deleted",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Subject"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Subject" : {
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "integer",
            "example" : 1
          },
          "Name" : {
            "type" : "string",
            "example" : "СТО"
          },
          "LectureHours" : {
            "type" : "integer",
            "example" : 150
          },
          "PracticeHours" : {
            "type" : "integer",
            "example" : 100
          },
          "LabHours" : {
            "type" : "integer",
            "example" : 50
          }
        },
        "example" : {
          "PracticeHours" : 100,
          "LectureHours" : 150,
          "LabHours" : 50,
          "Id" : 1,
          "Name" : "СТО"
        }
      },
...
      "entityData_1" : {
        "type" : "object",
        "properties" : {
          "Name" : {
            "type" : "string",
            "example" : "СТО"
          },
          "LectureHours" : {
            "type" : "integer",
            "example" : 150
          },
          "PracticeHours" : {
            "type" : "integer",
            "example" : 100
          },
          "LabHours" : {
            "type" : "integer",
            "example" : 50
          }
        }
      },
...
    }
  }
}